Started on Global Coordinates

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/roadrunner/util/SplineConstants.java
@@ -5,10 +5,10 @@ public class SplineConstants {
     public double x;
     public double y;
 
-    public int heading;
+    public double heading;
     public int pauseTime;
 
-    public SplineConstants(double in_x, double in_y, int in_heading, int in_pauseTime) {
+    public SplineConstants(double in_x, double in_y, double in_heading, int in_pauseTime) {
         x = in_x;
         y = in_y;
         heading = in_heading;
